Use multi suite function to have each JMS test be in its own FATSuite
