+++ dev/com.ibm.ws.springboot.fat20_fat/bnd.bnd
@@ -13,6 +13,8 @@
 -include= ~../cnf/resources/bnd/bundle.props
 bVersion=1.0
 
+-sub: *.bnd
+
 fat.project: true
 
 fat.test.container.images: icr.io/ibm-messaging/mq:9.3.2.0-r2
