+++ dev/io.openliberty.springboot.fat30_fat/fat/src/com/ibm/ws/springboot/support/fat/suite/jms2/FATSuite.java
@@ -1,0 +1,32 @@
+/*******************************************************************************
+ * Copyright (c) 2018, 2025 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.springboot.support.fat.suite.jms2;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+import com.ibm.ws.springboot.support.fat.JmsSpringBootAppTests30;
+
+import componenttest.containers.TestContainerSuite;
+import componenttest.custom.junit.runner.AlwaysPassesTest;
+
+@RunWith(Suite.class)
+@SuiteClasses({
+                AlwaysPassesTest.class,
+                JmsSpringBootAppTests30.class
+})
+
+public class FATSuite extends TestContainerSuite {
+    // Empty
+}
