+++ dev/io.openliberty.springboot.fat40_fat/bnd.bnd
@@ -13,6 +13,8 @@
 -include= ~../cnf/resources/bnd/bundle.props
 bVersion=1.0
 
+-sub: *.bnd
+
 javac.source: 17
 javac.target: 17
