+++ dev/io.openliberty.springboot.fat40_fat/fat/src/com/ibm/ws/springboot/support/fat/suite/jms1/FATSuite.java
@@ -1,0 +1,31 @@
+/*******************************************************************************
+ * Copyright (c) 2018, 2025 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.springboot.support.fat.suite.jms1;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+import com.ibm.ws.springboot.support.fat.JmsWebAppTests40;
+
+import componenttest.containers.TestContainerSuite;
+import componenttest.custom.junit.runner.AlwaysPassesTest;
+
+@RunWith(Suite.class)
+@SuiteClasses({
+                AlwaysPassesTest.class,
+                JmsWebAppTests40.class
+})
+
+public class FATSuite extends TestContainerSuite {
+}
