+++ dev/io.openliberty.springboot.fat40_fat/suite.jms2.bnd
@@ -1,0 +1,16 @@
+#*******************************************************************************
+# Copyright (c) 2018,2025 IBM Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License 2.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-2.0/
+# 
+# SPDX-License-Identifier: EPL-2.0
+#
+# Contributors:
+#     IBM Corporation - initial API and implementation
+#*******************************************************************************
+
+-privatepackage: com.ibm.ws.springboot.support.fat.suite.jms2
+
+suite.name: com.ibm.ws.springboot.support.fat.suite.jms2.FATSuite
